workshop/multi-embedding/requirements.txt
voyager==2.0.9
pylate
weaviate-client
sentence-transformers